# docs/swagger.yaml
basePath: /v1
definitions:
  controller.GetPowerState:
    properties:
      state:
        enum:
        - "ON"
        - "OFF"
        type: string
    type: object
  controller.ResBool:
    properties:
      message:
        type: string
      result:
        type: boolean
    type: object
  controller.ResError:
    properties:
      message:
        type: string
    type: object
  controller.ResValidationError:
    properties:
      message:
        type: string
      validationErrors:
        items:
          $ref: '#/definitions/controller.ResValidationErrorDetail'
        type: array
    type: object
  controller.ResValidationErrorDetail:
    properties:
      message:
        type: string
      param:
        type: string
    type: object
  controller.SetPowerState:
    properties:
      state:
        enum:
        - "ON"
        - "OFF"
        - TOGGLE
        type: string
    type: object
  models.Computer:
    properties:
      computer_id:
        type: integer
      ip_address:
        type: string
      mac_address:
        type: string
      name:
        type: string
      ssh_key:
        type: string
      ssh_port:
        type: number
      ssh_user:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: PCの電源をON/OFFするためのAPIです
  license:
    name: ISC
    url: https://licenses.opensource.jp/ISC/ISC.html
  title: PC Remote Switch API
  version: "1.0"
paths:
  /computers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Computer'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ResError'
      summary: コンピュータの一覧を取得
      tags:
      - computer
    post:
      consumes:
      - application/json
      parameters:
      - description: 'ssh_key: default $HOME/.ssh/id_rsa, ssh_port: default 22, mac_address:
          default ''arp -a ip_address'''
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Computer'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResValidationError'
      summary: コンピュータを登録
      tags:
      - computer
  /computers/{computerId}:
    delete:
      parameters:
      - description: Computer ID
        in: path
        name: computerId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ResError'
      summary: コンピュータを削除
      tags:
      - computer
    get:
      parameters:
      - description: Computer ID
        in: path
        name: computerId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Computer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ResError'
      summary: コンピュータの詳細を取得
      tags:
      - computer
    put:
      consumes:
      - application/json
      parameters:
      - description: Computer ID
        in: path
        name: computerId
        required: true
        type: integer
      - description: 'ssh_key: default $HOME/.ssh/id_rsa, ssh_port: default 22, mac_address:
          default ''arp -a ip_address'''
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Computer'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResValidationError'
      summary: コンピュータを更新
      tags:
      - computer
  /computers/{computerId}/power:
    get:
      parameters:
      - description: Computer ID
        in: path
        name: computerId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetPowerState'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ResError'
      summary: コンピュータの電源状態を取得
      tags:
      - power
    put:
      consumes:
      - application/json
      parameters:
      - description: Computer ID
        in: path
        name: computerId
        required: true
        type: integer
      - description: 変更したい電源状態
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.SetPowerState'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResBool'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ResError'
      summary: コンピュータの電源状態を変更
      tags:
      - power
swagger: "2.0"